main comments + changes to readme + SLIGHT changes to bbs
Sometimes BBS was generating duplicates and then there wouldn't be enough vertices in in the ArrayList of nodes (since our graph does not allow for duplicate vertex values), so I was getting an index out of bounds exception.
SOLUTION: I changed buffer() in the BBS class to return a HashSet (doesn't allow duplicates) instead of an ArrayList. Then, in main, I made it so that it keeps generating a new HashSet of node values until there are NO duplicates.

BBS with a bitNum of < 3 cannot generate 15 unique node values. I forced the user to enter a bitNum greater than 5 so it would eliminate the possibility of this happening.